@@ -24,11 +24,9 @@
   },
   "homepage": "https://github.com/Hokkaidosunny/electron-tunnel#readme",
   "devDependencies": {
-    "@types/electron": "^1.6.10",
+    "electron": "1.7.9",
     "tslint": "^5.8.0",
     "typescript": "^2.6.1"
   },
-  "dependencies": {
-    "electron": "^1.6.10"
-  }
+  "dependencies": {}
 }
